+++ AmigoPJT/AmigoPJT.vcxproj.filters
@@ -18,9 +18,6 @@
     <ClCompile Include="command.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="DelCommand.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
     <ClCompile Include="main.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
